refactor "custom" to "advanced" in the code
