+++ frontend/src/scenes/plugins/tabs/advanced/LocalPlugin.tsx
